pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sleepless-agent"
version = "0.1.2"
description = "24/7 AI agent that maximizes Claude usage via Slack interface"
readme = "README.md"
authors = [{name = "Sleepless Agent Maintainers"}]
license = {file = "LICENSE"}
keywords = ["ai", "automation", "agents", "slack", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business",
]
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "slack-sdk",
    "sqlalchemy",
    "python-dotenv",
    "PyYAML",
    "gitpython",
    "aiosqlite",
    "loguru",
    "rich",
    "structlog",
    "psutil",
    "claude-agent-sdk",
    "requests",
    "uvicorn[standard]",
]

[project.scripts]
sle = "sleepless_agent.__main__:main"

[project.urls]
Homepage = "https://github.com/context-machine-lab/sleepless-agent"
Repository = "https://github.com/context-machine-lab/sleepless-agent"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"sleepless_agent" = ["config.yaml", "interfaces/web_ui/static/*", "interfaces/web_ui/templates/*"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["sleepless_agent*"]
